[DRAFT]: Extract Push Model information from TPM

+++ keylime/src/structures/capabilities_negotiation.rs
@@ -64,8 +64,8 @@ pub struct Capabilities {
 #[derive(Serialize, Deserialize, Debug, Clone)]
 // Do not serialize the struct name, only the fields
 pub struct ShaValues {
-    pub sha1: Vec<u8>,
-    pub sha256: Vec<u8>,
+    pub sha1: Vec<u32>,
+    pub sha256: Vec<u32>,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
